// frontend/src/app/model/patient.ts
export class Patient {
  patientId: number = 0;
  patientTransectionId: number = 0;
  patientCardNo: string = '';
  patientName: string = '';
  //branchName:string='';
  patientBranchId: number = 0;
  branchName: string = '';
  patientDistrictName: string = '';
  patientStateName: string = '';
  patientDateOfUsage: Date = new Date();
}